splits tests per level upto 7
